refactor(diagnose): move drains_to_dprst diagnostics under scripts/diagnose/ (#147)
