+++ scripts/diagnose/__init__.py
@@ -1,0 +1,1 @@
+"""Per-VPU drains_to_dprst diagnostics (coverage + A/B harness)."""
