+++ scripts/diagnose/diagnose_drains_to_dprst.py
